계층별 계약과 기능 중심 구조 재정비
- Application.Contracts의 미디어·내비게이션·노트·쿼리 계약을 기능 단위 네임스페이스로 재배치
- Shared 프로젝트 제거 및 공용 타입의 Domain·Application.Contracts·Presentation 계층 귀속 정리
- Shell.Contracts의 Presentation.Contracts 전환과 프로젝트 참조 관계 단순화
- 설정 저장소 추상화와 애플리케이션·화면 상태 설정 서비스 분리
- 내비게이션·노트 서비스와 모델의 기능 중심 디렉터리 구조 통일
- 공용 명령·도우미·매퍼·메시지의 Presentation 프로젝트 통합
- DI 등록과 서비스 수명 및 관련 XAML 바인딩·네임스페이스 정비

+++ MyNotes/ViewModels/Notes/NoteEditorViewModel.cs
@@ -5,19 +5,18 @@ using CommunityToolkit.WinUI.Helpers;
 
 using Microsoft.Windows.Storage.Pickers;
 
-using MyNotes.Application.Commands.Notes;
-using MyNotes.Application.Contracts.Models.Notes;
+using MyNotes.Application.Contracts.Converters;
+using MyNotes.Application.Contracts.Media.Models;
+using MyNotes.Application.Contracts.Notes.Models;
+using MyNotes.Application.Notes.Commands;
+using MyNotes.Application.Notes.Services;
 using MyNotes.Application.Results;
-using MyNotes.Application.Services.Notes;
 using MyNotes.Common.Collections;
 using MyNotes.Common.Commands;
 using MyNotes.Common.Helpers;
+using MyNotes.Constants;
 using MyNotes.Models.Notes;
 using MyNotes.Services.Windows;
-using MyNotes.Shared.Constants;
-using MyNotes.Shared.Enums.Media;
-using MyNotes.Shared.Enums.Notes;
-using MyNotes.Shell.Contracts.Converters;
 using MyNotes.Templates.Media;
 
 namespace MyNotes.ViewModels.Notes;
@@ -177,7 +176,7 @@ internal sealed partial class NoteEditorViewModel : ViewModelBase, IAsyncDisposa
           PatchDto = new NoteViewStatePatchDto()
           {
             Id = Note.Id,
-            BackgroundImageAlignment = (int)Note.BackgroundImageAlignment
+            BackgroundImageAlignment = Note.BackgroundImageAlignment
           }
         });
         break;
@@ -199,7 +198,7 @@ internal sealed partial class NoteEditorViewModel : ViewModelBase, IAsyncDisposa
           PatchDto = new NoteViewStatePatchDto()
           {
             Id = Note.Id,
-            BackdropKind = (int)Note.BackdropKind
+            BackdropKind = Note.BackdropKind
           }
         });
         break;
